+++ packages/nx-plugin/src/open-api/ts-client/generator.utils.spec.ts
@@ -6,6 +6,10 @@ import { Tree } from '@nx/devkit';
 import { createProjectSync } from '@ts-morph/bootstrap';
 import ts from 'typescript';
 import { createTreeUsingTsSolutionSetup } from '../../utils/test';
+import { Mock } from 'vitest';
+import { importTypeScriptModule } from '../../utils/js';
+
+export const baseUrl = 'https://example.com';
 
 export const expectTypeScriptToCompile = (
   tree: Tree,
@@ -37,6 +41,64 @@ export const expectTypeScriptToCompile = (
     console.log(project.formatDiagnosticsWithColorAndContext(diagnostics));
   }
   expect(diagnostics).toHaveLength(0);
+};
+
+export const callGeneratedClient = async (
+  clientModule: string,
+  mockFetch: Mock<any, any>,
+  op: string,
+  parameters?: any,
+): Promise<any> => {
+  const { TestApi } = await importTypeScriptModule<any>(clientModule);
+  const client = new TestApi({ url: baseUrl, fetch: mockFetch });
+  const clientMethod = op.split('.').reduce((m, opPart) => m[opPart], client);
+  return await clientMethod(parameters);
+};
+
+export const callGeneratedClientStreaming = async (
+  clientModule: string,
+  mockFetch: Mock<any, any>,
+  op: string,
+  parameters?: any,
+): Promise<AsyncIterableIterator<any>> => {
+  const { TestApi } = await importTypeScriptModule<any>(clientModule);
+  const client = new TestApi({ url: baseUrl, fetch: mockFetch });
+  const clientMethod = op.split('.').reduce((m, opPart) => m[opPart], client);
+  return clientMethod(parameters);
+};
+
+export const mockStreamingFetch = (
+  status: number,
+  chunks: any[],
+): Mock<any, any> => {
+  const mockFetch = vi.fn();
+
+  let i = 0;
+
+  const mockReader = vi.fn();
+  mockReader.mockReturnValue({
+    read: vi.fn().mockImplementation(() => {
+      const value = chunks[i];
+      const done = i >= chunks.length;
+      i++;
+      return {
+        done,
+        value,
+      };
+    }),
+  });
+
+  mockFetch.mockResolvedValue({
+    status,
+    body: {
+      pipeThrough: () => ({
+        getReader: mockReader,
+      }),
+      getReader: () => mockReader,
+    },
+  });
+
+  return mockFetch;
 };
 
 // A couple of tests for the test utility as a sanity check
